iris/backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
scikit-learn==1.3.0
psycopg2-binary==2.9.7
pandas==2.0.3
